Fixed (very) tiny memory leak. Make explicit typecasts (to prevent compiler warnings). Return success indication from CIGet. Support unpacking bool arrays. Use array delete when cleaning up message buffers.

+++ Model/ComponentInterface2/CMPComponentEntryPoints.cpp
@@ -33,6 +33,6 @@ extern "C" void EXPORT STDCALL getDescriptionLength(char* initScript, int* lengt
    {
    char* buffer = new char[500000];
    getDescriptionInternal(initScript, buffer);
-   *length = strlen(buffer);
+   *length = (int)strlen(buffer);
    delete [] buffer;
    }
